separate LAhE/NAhE rules in operand-3.
This change is consistent with operand-start, which doesn't group
these rules together.

This change is consistent with all other rules in operand-3, which
don't try to share common-substructure.

This change moves an ordered choice from the inside one choice out
to the loop with the other choices.  Since ordered choice is
something that can be checked in parallel, this change allows more
simultaneous threads to try parsing operand-3 at once.