simplify statement-2 by reducing two choices into one.
The two choices in statement-2 differ only by the final element in
the outermost optional clause.  The first choice does not have
statement-2, the second choice does.  This can be expressed by
another '?' operator, removing one ordered choice.

If this rule was constructed this way in order to avoid a bug in
the RATS! parser, please ignore this change.