fold li-clause into sumti-6.
This is the only choice in sumti-6 separated into it's own
production.  I'm not sure why it is pulled out, it is only
referenced in sumti-6.

If you were doing this because you wanted the parse tree
shaped a particular way, please ignore this change.

Otherwise, this change is more consistent with the rest of
the grammar.